chore: Clean up `pg_analytics` tests (#969)
