+++ pg_analytics/src/types/timestamp.rs
@@ -35,8 +35,6 @@ impl TryFrom<PgTypeMod> for PgTimestampPrecision {
 
         match typemod {
             -1 => Ok(PgTimestampPrecision::Default),
-            0 => Ok(PgTimestampPrecision::Second),
-            3 => Ok(PgTimestampPrecision::Millisecond),
             6 => Ok(PgTimestampPrecision::Microsecond),
             unsupported => Err(TimestampError::UnsupportedTypeMod(unsupported)),
         }
